Adds team_statistics module for method grouping. Corrects method naming to mirror spec.
